后台数据能拿到，未解决$ JQUERY不识别的问题

+++ src/main/java/com/lena/controller/PermissionController.java
@@ -39,7 +39,7 @@ public class PermissionController {
 
     @RequestMapping(value="/menu",method=RequestMethod.GET)
     @ResponseBody
-    public Results getmenu(Long userId){
+    public Results getmenu(Integer userId){
         return permissionService.getMenu(userId);
     }
